Archive Maintenance by fixGOsvn@22 * drop optional php-end-tag at end of php and inc files
according to http://us3.php.net/basic-syntax.instruction-separation the
closing tag of a PHP block at the end of a file is optional and includes
an immediately trailing newline if present, but no other spaces nor newlines.
Moreover any extra whitespace at the end of the file means trouble,
hence the guideline to omit such a closing tag at the end of a file.

see http://lists.drupal.org/archives/development/2005-08/msg00648.html
for a long discussion of this topic on the drupal mailing list